refactor(mobile): restore the JavaScript tab bar
