--- apps/mobile/src/app/(app)/(tabs)/AndroidTabsLayout.tsx
@@ -1,77 +1,0 @@
-import { useSafeAreaInsets } from "react-native-safe-area-context";
-import { Tabs } from "expo-router";
-import { useTranslation } from "react-i18next";
-import { useTheme } from "styled-components/native";
-
-import Logo from "@/assets/images/Logo";
-import Messages from "@/assets/images/Messages";
-import Profile from "@/assets/images/Profile";
-
-interface TabBarIconProps {
-  focused: boolean;
-  color: string;
-}
-
-/**
- * JS-rendered tabs, kept on Android. See `_layout.tsx` for why Android
- * doesn't get Native Tabs in this pass.
- */
-export default () => {
-  const theme = useTheme();
-  const insets = useSafeAreaInsets();
-  const { t } = useTranslation();
-
-  return (
-    <Tabs
-      screenOptions={{
-        sceneStyle: { backgroundColor: theme.colors.background },
-        tabBarInactiveTintColor: theme.colors.text,
-        tabBarActiveTintColor: theme.colors.primary,
-        tabBarShowLabel: false,
-        headerShown: false,
-        tabBarStyle: {
-          backgroundColor: theme.colors.background,
-          borderTopWidth: 0,
-          elevation: 0,
-          marginVertical: theme.spacing[3],
-          marginBottom: insets.bottom ? theme.spacing[0.5] : theme.spacing[3],
-        },
-        tabBarBadgeStyle: {
-          backgroundColor: theme.colors.primary,
-          color: theme.colors.background,
-        },
-      }}
-    >
-      <Tabs.Screen
-        name="swipe"
-        options={{
-          title: t("tabs.swipe"),
-          tabBarButtonTestID: "tab-swipe",
-          tabBarIcon: ({ color }: TabBarIconProps) => (
-            <Logo colorStopOne={color} colorStopTwo={color} width={34} height={34} />
-          ),
-        }}
-      />
-      <Tabs.Screen
-        name="messages"
-        options={{
-          title: t("tabs.messages"),
-          tabBarButtonTestID: "tab-messages",
-          tabBarIcon: ({ color }: TabBarIconProps) => (
-            <Messages colorStopOne={color} colorStopTwo={color} width={34} height={34} />
-          ),
-        }}
-      />
-      <Tabs.Screen
-        name="profile"
-        options={{
-          title: t("tabs.profile"),
-          tabBarButtonTestID: "tab-profile",
-          tabBarIcon: ({ color }: TabBarIconProps) => (
-            <Profile colorStopOne={color} colorStopTwo={color} width={34} height={34} />
-          ),
-        }}
-      />
-    </Tabs>
-  );
-};
