+++ packages/shared/i18n/locales/en/translation.json
@@ -327,10 +327,5 @@
     "automatic": "Automatic",
     "dark": "Dark",
     "light": "Light"
-  },
-  "tabs": {
-    "swipe": "Swipe",
-    "messages": "Messages",
-    "profile": "Profile"
   }
 }
