+++ packages/shared/i18n/locales/pt-BR/translation.json
@@ -327,10 +327,5 @@
     "automatic": "Automático",
     "dark": "Escuro",
     "light": "Claro"
-  },
-  "tabs": {
-    "swipe": "Descobrir",
-    "messages": "Mensagens",
-    "profile": "Perfil"
   }
 }
